Added matrix inverse and matrix reduced-row-echelon functions.
At this point in the course we have been tasked with developing our own function library (SPmath) for single-precision matrices. This is my initial step to solve this problem.

The standard algorithm with largest pivots method (i.e: wikipedia) was added for determination of the reduced-row-echelon form of matrices of any side.
An inverse function was added for square matrices.
A determinant function remains to be added.
A linear solver remains to be added.