Added rank and determinant calculation to SPmath
rankm(): computes the rank of a matrix based on forward elimination

detm(): computes the determinant of a matrix based on row operations in forward elimination

In addition, a few optimizations have been added to the rest of the functions to improve stability. Mostly changing ints to unsigned ints where they are the size of objects.

Relevant to Issue #1.