Added invertible check and runtime error throw to matrix inversion.
A check was added to invertm() to determine if, after Gauss-Jordan elimination, the left hand side of the composite matrix has any zeros in the diagonal.

This is, of course, a tad inefficient, as it adds overhead to cases where the matrix is non invertible, but in those cases, performance is usually not the limiting issue. IMHO it is better to limit the peformance of the non-invertible cases with a post-hoc test than hurt the invertible cases with a costly check before calling or increase complexity by restructuring the library to include a test partway through inversion (which would be optimal.)

Related to Issue #1.