Fixed bug writing pin states that caused multiple to change.
There was a bug that caused multiple pins to blink when writing output to certain pins. The issue was that I was using the same method as pin mode settings and writing the old state of the register out to the register along with the changed pin everytime. Silly mistake. Eliminated that and it works now.

Resolves issue #2.