Completed Lab 7.
Updated all code and bugfixes to complete Lab 7.

The current lab will scan the room with the lidar and output the point cloud to a .log file.

There are some issues with random instability. A few times, the PanTilt and Lidar have quit altogether on me until a full powercycle. I am not sure why. It works fine normally.